Quality of Life commit
Switched from using SmapReading catch-all to more specialized structs,
found in archiver/types.go. This patch simplifies much of the code and
removed a lot of nested number conversions. The handlers other than HTTP
will have to be fixed to use the new interface from the archiver (needs
a type switch), because the archiver no longer serializes to JSON